Move behavior from BalanceViewController to BalanceViewModel
