+++ CombineIntro/BalanceFeature/BalanceViewModel.swift
@@ -1,0 +1,44 @@
+import Combine
+import Foundation
+import UIKit
+
+final class BalanceViewModel {
+    private(set) var state = BalanceViewState()
+    private let service: BalanceService
+    private var cancellables: Set<AnyCancellable> = []
+
+    init(service: BalanceService) {
+        self.service = service
+
+        NotificationCenter.default
+            .publisher(for: UIApplication.willResignActiveNotification)
+            .sink { [weak self] _ in
+                self?.state.isRedacted = true
+            }
+            .store(in: &cancellables)
+
+        NotificationCenter.default
+            .publisher(for: UIApplication.didBecomeActiveNotification)
+            .sink { [weak self] _ in
+                self?.state.isRedacted = false
+            }
+            .store(in: &cancellables)
+    }
+
+    func refreshBalance() {
+        state.didFail = false
+        state.isRefreshing = true
+        service.refreshBalance { [weak self] result in
+            self?.handleResult(result)
+        }
+    }
+
+    private func handleResult(_ result: Result<BalanceResponse, Error>) {
+        state.isRefreshing = false
+        do {
+            state.lastResponse = try result.get()
+        } catch {
+            state.didFail = true
+        }
+    }
+}
